+++ src/rafi-dump/Main.cpp
@@ -78,7 +78,7 @@ void PrintFpRegNode(const FpRegNode* node)
         printf(
             "    f%-2d: { u64: 0x%016llx, f32: %e, f64: %e } // %s\n",
             i,
-            node->regs[i].u64.value,
+            static_cast<unsigned long long>(node->regs[i].u64.value),
             node->regs[i].f32.value,
             node->regs[i].f64.value,
             GetFpRegName(i)
@@ -197,38 +197,38 @@ void PrintIntReg64Node(const IntReg64Node* node)
         "    x30: 0x%016llx // t5\n"
         "    x31: 0x%016llx // t6\n"
         "  }\n",
-        node->regs[0],
-        node->regs[1],
-        node->regs[2],
-        node->regs[3],
-        node->regs[4],
-        node->regs[5],
-        node->regs[6],
-        node->regs[7],
-        node->regs[8],
-        node->regs[9],
-        node->regs[10],
-        node->regs[11],
-        node->regs[12],
-        node->regs[13],
-        node->regs[14],
-        node->regs[15],
-        node->regs[16],
-        node->regs[17],
-        node->regs[18],
-        node->regs[19],
-        node->regs[20],
-        node->regs[21],
-        node->regs[22],
-        node->regs[23],
-        node->regs[24],
-        node->regs[25],
-        node->regs[26],
-        node->regs[27],
-        node->regs[28],
-        node->regs[29],
-        node->regs[30],
-        node->regs[31]
+        static_cast<unsigned long long>(node->regs[0]),
+        static_cast<unsigned long long>(node->regs[1]),
+        static_cast<unsigned long long>(node->regs[2]),
+        static_cast<unsigned long long>(node->regs[3]),
+        static_cast<unsigned long long>(node->regs[4]),
+        static_cast<unsigned long long>(node->regs[5]),
+        static_cast<unsigned long long>(node->regs[6]),
+        static_cast<unsigned long long>(node->regs[7]),
+        static_cast<unsigned long long>(node->regs[8]),
+        static_cast<unsigned long long>(node->regs[9]),
+        static_cast<unsigned long long>(node->regs[10]),
+        static_cast<unsigned long long>(node->regs[11]),
+        static_cast<unsigned long long>(node->regs[12]),
+        static_cast<unsigned long long>(node->regs[13]),
+        static_cast<unsigned long long>(node->regs[14]),
+        static_cast<unsigned long long>(node->regs[15]),
+        static_cast<unsigned long long>(node->regs[16]),
+        static_cast<unsigned long long>(node->regs[17]),
+        static_cast<unsigned long long>(node->regs[18]),
+        static_cast<unsigned long long>(node->regs[19]),
+        static_cast<unsigned long long>(node->regs[20]),
+        static_cast<unsigned long long>(node->regs[21]),
+        static_cast<unsigned long long>(node->regs[22]),
+        static_cast<unsigned long long>(node->regs[23]),
+        static_cast<unsigned long long>(node->regs[24]),
+        static_cast<unsigned long long>(node->regs[25]),
+        static_cast<unsigned long long>(node->regs[26]),
+        static_cast<unsigned long long>(node->regs[27]),
+        static_cast<unsigned long long>(node->regs[28]),
+        static_cast<unsigned long long>(node->regs[29]),
+        static_cast<unsigned long long>(node->regs[30]),
+        static_cast<unsigned long long>(node->regs[31])
     );
 }
 
@@ -251,8 +251,8 @@ void PrintPc64Node(const Pc64Node* node)
         "    vaddr: 0x%016llx\n"
         "    paddr: 0x%016llx\n"
         "  }\n",
-        node->virtualPc,
-        node->physicalPc
+        static_cast<unsigned long long>(node->virtualPc),
+        static_cast<unsigned long long>(node->physicalPc)
     );
 }
 
@@ -288,7 +288,7 @@ void PrintTrap64Node(const Trap64Node* node)
         GetString(node->from),
         GetString(node->to),
         GetCauseString(node->cause),
-        node->trapValue
+        static_cast<unsigned long long>(node->trapValue)
     );
 }
 
@@ -304,9 +304,9 @@ void PrintMemoryAccessNode(const MemoryAccessNode* node)
         "  }\n",
         GetString(node->accessType),
         node->size,
-        node->value,
-        node->virtualAddress,
-        node->physicalAddress
+        static_cast<unsigned long long>(node->value),
+        static_cast<unsigned long long>(node->virtualAddress),
+        static_cast<unsigned long long>(node->physicalAddress)
     );
 }
 
@@ -330,7 +330,7 @@ void PrintCsr64Node(const Csr64Node* pNodes, int64_t nodeCount)
     for (int64_t i = 0; i < nodeCount; i++)
     {
         const auto address = static_cast<csr_addr_t>(pNodes[i].address);
-        printf("%16s: 0x%016llx\n", GetString(address), pNodes[i].value);
+        printf("%16s: 0x%016llx\n", GetString(address), static_cast<unsigned long long>(pNodes[i].value));
     }
 
     printf("  }\n");
